test(muya): extract shared bootMuya harness to cut duplication (Sonar)
SonarCloud quality gate flagged new-code duplicated-lines at 5.4% (threshold
3%), concentrated in the fold test files that each re-declared the same
bootMuya + MUYA_VERSION setup/teardown.

Extract that boilerplate into src/__tests__/muyaHarness.ts (useMuyaHarness +
bootMuya) and consume it from headingFold.spec.ts and foldActions.spec.ts. This
removes the duplicated setup blocks (the real duplication) rather than gaming
the metric, and future block/UI specs can reuse it.

Verified: muya tsc clean; lint clean; full muya suite 1508 tests pass (fold
suites 55).
